+++ src/alsyntaxmodifiers/iAddPageFieldCaptionsSettings.ts
@@ -1,0 +1,8 @@
+export interface IAddPageFieldCaptionsSettings {
+    setActionsCaptions: boolean | undefined;
+    setActionGroupsCaptions: boolean | undefined;
+    setGroupsCaptions: boolean | undefined;
+    setPartsCaptions: boolean | undefined;
+    setFieldsCaptions: boolean | undefined;
+    setLabelsCaptions: boolean | undefined;
+}